Now cases where there are no unapplied operations, like a sum where it is just a product with no extra additions, are taken into account. ie.:
(y * 4) ---> t1:=1*4

In the example not only is it not the sum of multiple numbers. It also does not start as the product of multiple numbers. It is the product of multiple numbers, yes, but hidden inside a math operator called a group. Because of the parentheses.

And so, the parsing goes like this: It starts by wanting to parse the unapplied sums. It does not found any, so it proceeds to parse the product. It wants to find multiple numbers multiplying themselves. It does not, instead it finds a single mathematical factor. It proceeds to parse it, finds out it is a group. Parses what's inside, and again, not the sum on multiple things. Just one. Proceeds to parse the product then. Finds the product of 1 and 4. Finishes.