Okay. So I remember just an hour ago saying that using bigger information in the parameters would help solve the problem. That was a good step, in the right direction.
Now, what I actually did was to implement each of the four cases independently inside the parseSum function.

case 1: there is a sum of two non-groups.
case 2: there is a sum of a left group and a right final
case 3: there is a sum of a left final and a right group
case 4: there is a sum of a left group and a right group

This allowed to turn this expression:
9 + ((1 * 3) + (2* 3)) + 5

into this:

t1:=2*3
t2:=1*3
t3:=t2+t1
t4:=9+t3
t5:=9+5