Started migration to Scala.
Why did I migrate to Scala:

There are four combinations of math operands in a monoid. Be that monoid a sum, a product, or a factor.

They were explained a few commits ago.

The thing is that these 3 monoids share the general algorithm for parsing the four said cases.

And I do not want to repeat myself.

So, I started by using functions as parameters. Be it parseProduct, or parseFactor, they all would get passed as parameter for the generic function we will call parse.

So, parse receives those functions and executes them. Right. But there was a problem. Those functions had different signatures. parseProduct would expect a ProductContext, while parseFactor a FactorContext. No problem, right? I would make them share a more generic signature, that would expect an object instead.

That was already a bad smell.

But the worse came during the moment to cast this objects to ProductContext or FactorContext. It was a mess. So, something smarter was much needed. A class that would be either of those.

And that cannot be done without going to the procedurally generated code made by ANTLR. A minimal change on the grammar would override the changes. And for this problem, Scala was needed.